+++ CHANGELOG.md
@@ -1,9 +1,25 @@
 # Change Log
 
-## [7.15.0](https://github.com/plivo/plivo-go/tree/v7.15.0) (2022-10-27)
+## [7.19.0](https://github.com/plivo/plivo-go/tree/v7.19.0) (2022-02-10)
 **Audio Streaming**
 - API support for starting, deleting, getting streams on a live call
 - XML creation support for stream element
+
+## [7.18.0](https://github.com/plivo/plivo-go/tree/v7.18.0) (2023-01-25)
+**Feature - Added New Param 'requester_ip' in Get Message and List Mssage APIs**
+- Add `requester_ip` to the response for the [list all messages API](https://www.plivo.com/docs/sms/api/message/list-all-messages/) and the [get message details API](https://www.plivo.com/docs/sms/api/message#retrieve-a-message)
+
+## [7.17.1](https://github.com/plivo/plivo-go/tree/v7.17.1) (2023-01-18)
+**Feature - Adding new param 'message_expiry' in Send Message API**
+-  Added new param 'message_expiry' in Send Message API 
+
+## [7.16.0](https://github.com/plivo/plivo-go/tree/v7.16.0) (2022-12-16)
+**Feature - Update campaign**
+- Update campaign API
+
+## [7.15.0](https://github.com/plivo/plivo-go/tree/v7.15.0) (2022-12-06)
+**Feature - Delete campaign and brand API**
+- Added Delete campaign and brand API
 
 ## [7.14.0](https://github.com/plivo/plivo-go/tree/v7.14.0) (2022-10-17)
 **Feature - Brandusecase API, 10DLC api enhancements**